Revert "dedup: read query files directly with QueryReader"
This reverts commit 97ae4d9d41cd440a6cdf0dee758e8f3b36b1efd6.
