+++ utils/src/bin/zpaste.rs
@@ -1,9 +1,9 @@
 use std::env;
-use monotextor_utils::queryreader::QueryReader;
+use monotextor_utils::ZPaste;
 
 fn main() {
     let args: Vec<String> = env::args().skip(1).collect();
-    let mf = QueryReader::new(args);
+    let mf = ZPaste::new(args);
     for i in mf {
         println!("{}", i);
     }
