+++ utils/src/lib.rs
@@ -1,12 +1,13 @@
+use std::io::{BufRead, BufReader, Lines};
+use std::fs::File;
 use serde::{Deserialize, Serialize};
 use gaoya::minhash::{MinHasher32, MinHasher};
 use gaoya::text::whitespace_split;
+use zstd::stream::read::Decoder;
 use shingles::Shingles;
 use fnv::FnvBuildHasher;
 use clap::ArgEnum;
 use seahash;
-
-pub mod queryreader;
 
 pub trait TextField {
     fn get_text(&self) -> String;
@@ -158,5 +159,68 @@ impl MinHashProcessor {
                     Shingles::new_with_step(text, self.window_size, self.window_size))
             }
         }
+    }
+}
+
+
+// An alternate implementation of unix 'paste' command
+// that reads and concatenates the lines of compressed files.
+// Store opened file descriptors of compressed files,
+// string buffer and specify field separator.
+pub struct ZPaste<'a> {
+    // quite a long generic specification, is there a simpler way?
+    readers: Vec<Lines<BufReader<Decoder<'a, BufReader<File>>>>>,
+    buf: String,
+    separator: char,
+}
+
+impl<'a> ZPaste<'a> {
+    pub fn new(names: Vec<String>) -> ZPaste<'a> {
+        let mut readers:
+            Vec<Lines<BufReader<Decoder<BufReader<File>>>>>
+            = Vec::new();
+
+        // Open each file and add to the vec its descriptor
+        for n in names {
+            let file = File::open(n).unwrap();
+            let decoder = Decoder::new(file).unwrap();
+            readers.push(BufReader::new(decoder).lines());
+        }
+
+        Self {
+            readers: readers,
+            buf: String::with_capacity(100),
+            separator: '\t',
+        }
+    }
+}
+
+impl Iterator for ZPaste<'_> {
+    type Item = String;
+
+    // Get one line, concatenation of all the lines in the files
+    fn next(&mut self) -> Option<String>{
+        self.buf.clear();
+        let mut readed = false;
+        let size = self.readers.len();
+
+        // Read one line of each file, concat all of them with separator
+        for (i, reader) in self.readers.iter_mut().enumerate() {
+            if let Some(line) = reader.next() {
+                self.buf.push_str(&line.unwrap());
+                readed = true;
+            }
+            if i != (size - 1) {
+                // Insert separator also for empty lines
+                self.buf.push(self.separator);
+            }
+        }
+        if !readed {
+            // If none of the readers provided data, stop
+            return None;
+        }
+
+        // clone only allocates needed memory, not all String capacity, so we safe
+        Some(self.buf.clone())
     }
 }
